requirements.txt
numpy==1.23.5
scipy==1.10.1
pandas==2.1.4
matplotlib
torch==2.0.1
pytorchvideo==0.1.5
av==12.0.0
pillow==9.5.0
accelerate==0.27.2
tqdm
scikit-image==0.21.0
PyYAML
nd2
opencv-python
timm==0.9.7
joblib
pretrainedmodels==0.7.4
pytest
scikit-learn==1.5.0
seaborn>=0.13.0
einops==0.7.0
fastparquet
pyarrow
